Added conservation of energy
Some of the behaviours (those in saves _nofood_x with no _defence) took advantage of a sort of bug that allowed snakes to opperate whily efectively spending zero energy. The clusters and highways that formed were the way they managed to take advantage of this.

These strategies no longer work, and do not evolved. The saves of them have been preserved with _coed.